parser: Cleanup nginx-combined with splitFields()
